test: add trait of provider to short test class
